requirements.txt
pdf2image==1.16.3
Pillow==9.5.0
pypdf==3.12.1
streamlit==1.24.1
poppler-utils==0.1.0
openai
requests
langchain
langchain_core
langchain-community
langchain-openai